Add SET_LED_HSI g-code
This is a g-code, due to the need to be aware of the colorspace
to determine whether to use the third pixel to fake white
for desaturation, or to actually use white. If PrinterNeoPixel
was exposed to the jinja context, this could be done as a macro.

I used the hsi to rgb conversion as published by SaikoLED for
use in their RGBW LED bulbs, found at
https://blog.saikoled.com/post/44677718712

  Signed-off-by: Franklyn Tackitt <im@frank.af>